Create SPARQL query to test for invalid IDs
PROBLEM:
oboInOwl:id are manually entered in Protege and can be different
than the assigned class number, potentially lead to errors.

Actual Example:
In v2021-07-29 release Klippel-Feil Syndrome 3 (class DOID_0080591)
has an non-matching ID DOID:0080592.

FIX:
This SPARQL query identifies classes with oboInOwl:id not matching
the expected class and could be instituted as a release test.

Details:
This query is designed to work for all purl obo classes. It is
not limited to testing DOIDs only but should also identify such
errors in ontology imports. It  could be broadened further to any
entity that has an oboInOwl:id attribute, if desired.