cleanup reasoning with uberon, bfo, and ro
This commit fixes and issue where loading doid with uberon, bfo, and
ro (as we do in the NIF-Ontology) produces reasoning errors. The owl
below can reproduce the issue at this time.

```ttl
@prefix : <file:///ERROR/EMPTY/PREFIX/BANNED/> .
@prefix dc: <http://purl.org/dc/elements/1.1/> .
@prefix owl: <http://www.w3.org/2002/07/owl#> .
@prefix skos: <http://www.w3.org/2004/02/skos/core#> .

<:test.ttl> a owl:Ontology ;
    owl:imports <http://purl.obolibrary.org/obo/doid.owl>,
        <http://purl.obolibrary.org/obo/uberon.owl> ,
        <http://purl.obolibrary.org/obo/ro.owl>,
        <http://purl.obolibrary.org/obo/bfo.owl> .

```

There are to three object properties that need to be replaced with
disease oriented versions to deconflate the disease from incidence and
pathology.

For the time being I have replaced the properties with variants that
end in `#disease` however new properties should be requested in ro.

RO:0002220 adjacent to -> 'disease adjacent to'
implies domain is ind cont rather than spec dep cont fix is the same
as for location, using the current relation conflates disease with
pathology

RO:0002488 existence starts during -> 'incidence starts during'
implies that domain is independent continuant, using the current
relation conflates disease with incidence

RO:0002241 has exposure stressor -> 'incidence has exposure stressor'
implies domain is process but diseases are modelled as dispositions
process and disposition are disjoint creating a new property that
deconflates diseases (disposition) from their incidence (process),
using the current relation conflates disease with incidence

There is an incorrect pattern of usage that conflates feet and hands
(and similar cases). This is due to the fact that a restriction of an
intersection does not commute with an intersection of a restriction.

In this case location of disposition is (food and hand) should be
location of disposition is foot and location of disposition is hand
aka subClassOf (disease has location some foot and disease has
location some hand).