Require proper format of def src xrefs to protect loader
PROBLEM:
A recent error occurred where an ECO code was listed with
hasDbXref that caused the loader to fail.

TROUBLESHOOTING:
- Loader expects definition source xrefs to be identified by
oboInOwl:hasDbXref, to start with "url:" and be
a valid web address including the scheme (http or https).
- It does have some error handling when "url:" is missing. It
identifies valid web addresses and adds the "url:" prefix but
these corrections should happend in the doid-edit.owl file itself
and not in the loader.