Generalize rule to update slims (+ DO_rare_slim)
Slims to be automatically updated in doid-edit.owl are now
identified by SPARQL queries in the src/sparql/update folder as
those ending with "_slim.rq".

New SPARQL queries for updating slims should be named to exactly
match the slim name and should output the class IRIs to add to
the slim, excluding classes already in the slim.