Remove incorrect @fr lang tag
DOID:3390 has the exact synonym “Keratosis palmaris et plantaris”
which is tagged as being French, but it appears to be Latin. The
french translation is “Kératose palmaire et plantaire” according
to google translate (from English or Latin). Since this synonym
does appear to be used in English, it was decided that the @fr
lang tag should be removed and the synonym treated as English.