Update lang-rt.awk & make rules to add ontology annotations
This works finally.

NOTE on unusual multi-line output:
The awk script outputs each element of each --language-annotation
statement on its own line because each element needs to be
individually quoted to avoid shell quoting of special characters
when passed to the ROBOT command.

This is most noticeable when input text includes spaces. To
create a statement on one line requires quoting the text input, e.g.

--language-annotation dc11:description "The Disease Ontology..." en

When on a single line, the shell will quote all double quotes
breaking up this text into multiple elements/commands for robot, e.g.

--language-annotation dc11:description '"The' Disease 'Ontology..."' en

In the example, the single text input becomes 3 inputs with '"The'
treated as the text input, Disease as the language (totally wrong),
and ROBOT then thinks 'Ontology..."' is a new argument or subcommand.

*** This is only apparent when using set -x before the command, as
printing on the commandline does not show the single quoting. ***

NOTE on languages:
This is not intended for use with English since the annotation
would be added with an "en" tag, instead of without a tag as
preferred. This should not become an issue as long as languages
are hard-coded in the Makefile.