Move version_imports before products rule -- necessary
version_imports was moved after products in commit 23283e05
because it was linked with version_edit, and version_edit
causes the full release pipeline to re-run if there are any
issues since it was updated every time.

However, it is necessary for version_imports to be before the
products rule for releases because .json products include the
versionIRI of imports, and the wrong version will be captured if
the imports are not updated before the .json files are built.
Since the OWL products are reliant on ontology IRIs, they are not
affected.