Update Ebola virus disease
Renamed to Ebola virus disease following WHO and most common
use.

Additional notes:
- The most common name in publications is "Ebola", though it is
used to refer to both the virus and the disease and it is
difficult to differentiate which it is used for. "Ebola" alone
is not used by official government sources or WHO. It's been
added as a synonym.
- The next most common is "Ebola virus disease" (used by WHO),
then the current name in DO "Ebola hemorrhagic fever", then
"Ebola disease" (used by CDC). The CDC seems to prefer "Ebola
disease" and uses "Ebola virus disease" for the disease caused
specifically by Orthoebolavirus zairense. It makes the most sense
to use the prevailing name (most common currently, preferred by
official source), over the historical name, and to ensure all of
these are synonyms.

ICTV name revision, #1429