Change scope of Fanconi-Bickel syndrome mapping to MESH
This MESH ID is an exact mapping to both Fanconi syndrome and
Fanconi-Bickel syndrome, which is an error.

The MESH concept is Fanconi Syndrome and has Fanconi-Bickel
syndrome as an attached term. That is likely incorrect as these
appear to be considered distinct disease entities. Since users
may use this MESH ID for Fanconi-Bickel syndrome, the mapping has
been retained as an xref and skos:relatedMatch.